Time: 12 ms (44.63%), Space: 11.3 MB (70.01%) - LeetHub

+++ 990-satisfiability-of-equality-equations/990-satisfiability-of-equality-equations.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     
-    vector<int> parent;
+    vector<short int> parent;
     
-    int findd (int x) {
+    int findd (short int x) {
         if (x == parent[x]) return x;
         return findd(parent[x]);
     }
     
-    void unionn (int x, int y) {
+    void unionn (short int x, short int y) {
         int xP = findd(x);
         int yP = findd(y);
         parent[xP] = yP;
